Reduce many constructs to more Kotlin like code
Improvements in code on either Kotlin coding style and also algorithmic
approach. For example to test whether every driver knows every gossip,
it is enough to ask whether every driver knows exactly as many gossips
as there are drivers - compared to formerly testing for equality of the
sets of gossip.