Remove the concept of Gossip
To solve the kata, it is unnecessary to model the concept of a single
gossip as its own entity. Each driver only needs to know, whose other
gossips he has heard of - directly or indirectly.
Thus, a Driver carries a set of other Driver objects representing the
set of Drivers whose gossips have been transfered to this Driver.